Move queries in to app/queries and remove module Query
