--- app/models/query/people_never_logged_in_query.rb
@@ -1,17 +1,0 @@
-module Query
-
-  class PeopleNeverLoggedInQuery
-
-    class << self
-      delegate :call, to: :new
-    end
-
-    def initialize(relation = Person.all)
-      @relation = relation
-    end
-
-    def call
-      @relation.where(login_count: 0)
-    end
-  end
-end
